requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
db-dtypes==1.3.1
google-cloud-bigquery
google-cloud-storage
idna==3.10
numpy
packaging==24.2
pandas
pyarrow
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
six==1.16.0
tzdata==2024.2
urllib3~=1.26.20
boto3~=1.35.63
pip~=24.2
wheel~=0.41.2
protobuf~=5.28.1
cryptography
setuptools
cffi~=1.17.1
google-crc32c~=1.6.0